+++ popup-bom/popup-parent/popup-starter/popup-starter-elasticsearch/popup-starter-elasticsearch-autoconfigure/src/main/java/com/photowey/popup/starter/elasticsearch/autoconfigure/config/PopupElasticsearchAutoConfigure.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright © 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.photowey.popup.starter.elasticsearch.autoconfigure.config;
+
+import com.photowey.popup.starter.elasticsearch.autoconfigure.builder.DefaultElasticsearchQueryBuilder;
+import com.photowey.popup.starter.elasticsearch.autoconfigure.builder.ElasticsearchQueryBuilder;
+import com.photowey.popup.starter.elasticsearch.autoconfigure.operator.DefaultElasticsearchOperator;
+import com.photowey.popup.starter.elasticsearch.autoconfigure.operator.ElasticsearchOperator;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.data.elasticsearch.ElasticsearchDataAutoConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+
+/**
+ * {@code PopupElasticsearchAutoConfigure}
+ *
+ * @author photowey
+ * @date 2023/03/05
+ * @since 1.0.0
+ */
+@AutoConfiguration(after = ElasticsearchDataAutoConfiguration.class)
+@Import(value = {RestClientElasticsearchConfigure.class})
+public class PopupElasticsearchAutoConfigure {
+
+    @Bean
+    @ConditionalOnMissingBean(ElasticsearchQueryBuilder.class)
+    public ElasticsearchQueryBuilder elasticsearchQueryBuilder() {
+        // TODO
+        return new DefaultElasticsearchQueryBuilder();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean(ElasticsearchOperator.class)
+    public ElasticsearchOperator elasticsearchOperator() {
+        // TODO
+        return new DefaultElasticsearchOperator();
+    }
+}
