+++ popup-bom/popup-parent/popup-starter/popup-starter-elasticsearch/popup-starter-elasticsearch-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.photowey.popup.starter.elasticsearch.config.PopupElasticsearchAutoConfigure
+com.photowey.popup.starter.elasticsearch.autoconfigure.config.PopupElasticsearchAutoConfigure
